requirements.txt
pyyaml==6.0.1
types-PyYAML==6.0.12
coverage==7.6.4
mypy-extensions==1.0.0
mypy==1.11.2
packaging==24.1
pip==24.3.1
py==1.11.0
pydocstyle==6.3.0
pytest-cov==5.0.0
pytest==7.1.2
ruff==0.7.1
setuptools==65.5.0
